add widget

+++ src/Orchard.Web/Modules/Skywalker.Webshop/ViewModels/UpdateShoppingCartItemViewModel.cs
@@ -7,7 +7,7 @@ namespace Skywalker.Webshop.ViewModels
 {
     public class UpdateShoppingCartItemViewModel
     {
-        public decimal ProductId { get; set; }
+        public int ProductId { get; set; }
         public bool IsRemoved { get; set; }
         public int Quantity { get; set; }
     }
